% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grammar.R
\name{create_grammar}
\alias{create_grammar}
\title{grammar}
\usage{
create_grammar(...)
}
\value{
Grammar object
}
\description{
grammar
}
\examples{
create_grammar(a = "<b> + 1, <c>(<b>)",
        b = "1 + <c>(3), 2",
        c = "log, exp, max")
}
