% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grammar_table.R
\name{create_grammar_table}
\alias{create_grammar_table}
\title{create_grammar_table}
\usage{
create_grammar_table(
  Grammar,
  max.depth = NULL,
  return_unrolled_grammar = FALSE
)
}
\arguments{
\item{Grammar}{}
}
\value{
Empty search table for given grammar
}
\description{
Initializes a search table for a given Grammar.
}
\examples{
new_gram <- grammar(a = "<b> + 1, <c>(<b>)",
                    b = "1 + <c>(3), 2",
                    c = "log, exp, max")
search.table(new_gram)

recursive_grammar <- grammar(a = "<a> + <b>, <c>(<a>)",
                             b = "1 + <c>(3), <b>",
                             c = "log, exp")
search.table(recursive_grammar, max.depth = 4)

}
