\name{fso.internal}
\alias{core.fso}
\alias{plotid}
\alias{hilight}
\alias{chullord}
\title{FSO Internal Functions}
\description{These functions establish several generic functions,
and are not intended to be called directly.  core.fso is the basic
fuzzy set (axis) construction algorithm called by all other 
routines as necessary.  The others establish plotting routines}
\keyword{internal}
