% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_function.R
\name{generate_function}
\alias{generate_function}
\title{generate_function}
\usage{
generate_function(Grammar, grammar_table)
}
\arguments{
\item{Grammar}{Grammar object produced by the function grammar}

\item{grammar_table}{search table produced by grammar_table and filled with integer values larger 0}

\item{max.depth}{Maximum recursive depth given as an integer.}
}
\value{
string of function call
}
\description{
Applies search table on a Grammar object
}
\examples{
test <- grammar(a = "<b> + 1, <c>(<b>)",
                b = "1 + <c>(3), 2",
                c = "log, exp, max")

test_table <- grammar_table(test)
test_table$value <- c(2, 1, 1, 1)

apply.search(test, test_table)

recursive_grammar <- grammar(a = "<a> + <b>, <c>(<a>)",
                             b = "1 + <c>(3), <b>",
                             c = "log, exp")
recursive_table <- grammar_table(recursive_grammar, max.depth = 4)
recursive_table$value <- 1

# This recursive table does not work without specifying the max.depth also in the apply.search function
# This is due to the fact that in comlicated recursive function, it is hard to infere the max depth from the search table
generate_function(recursive_grammar, recursive_table)

# Now it works
apply.search(recursive_grammar, recursive_table, max.depth = 4)


}
