% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar_sampler.R
\name{grammar_sampler}
\alias{grammar_sampler}
\title{grammar_sampler}
\usage{
grammar_sampler(
  Grammar,
  max.depth,
  n,
  save_feather = TRUE,
  parallel = TRUE,
  no_cores = NULL
)
}
\arguments{
\item{Grammar}{A Grammar Object, created by the grammar() function.}

\item{max.depth}{Maximum recursive depth used to create search Table.}

\item{n}{Number of functions to be created from Grammar.}

\item{save_feather}{TRUE/FALSE, saves a feather of the output list}

\item{parallel}{Should code be run parallel or not? TRUE or FALSE}

\item{no_cores}{The number of cores for parallel computation. If NULL than no_cores = detectCores() - 2, hence all but 2 cores.}
}
\value{
Returns a df with all sampled grammar vectors and their corresponding functions. Each row is 1 grammar vector and 1 resulting function (last element).
}
\description{
grammar_sampler
}
\examples{
g_v1 <- grammar(tf = "numeric * <eq> + numeric, <eq> + numeric, <eq> ",
                eq = "<fs>, <eq><op><fs>, <eq><op>numeric, <fs><op>numeric",
                fs = "<sp>,  <f>(<sp>), <sp><op><sp>, numeric",
                f = "exp, log",
                op = "+, -, *, /",
                sp = "slope, evi, sand, clay, elevation, hand, noise")
g_v1_table <- search.table(g_v1, max.depth = 3)
my_function_list <- par_grammar_sampler(n = 30000000,
                                        cfgram = g_v1,
                                        gram_table = g_v1_table
                                        max.depth = 5)
}
