% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_numerics.R
\name{simplify_numerics}
\alias{simplify_numerics}
\title{simplify_functions}
\usage{
simplify_numerics(
  functions,
  num_string = "numeric",
  no_cores = NULL,
  save = TRUE,
  file_name = NULL
)
}
\arguments{
\item{numeric_string}{}
}
\value{

}
\description{
simplify_functions
}
