% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_input.R
\name{variable_input}
\alias{variable_input}
\title{variable_input}
\usage{
variable_input(
  functions,
  variables,
  numbers,
  n_iter = 100,
  no_cores = NULL,
  seed = NULL,
  save = TRUE
)
}
\arguments{
\item{seed}{}
}
\value{

}
\description{
variable_input
}
