% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar_sampler.R
\name{grammar_sampler}
\alias{grammar_sampler}
\title{grammar_sampler}
\usage{
grammar_sampler(
  n,
  grammar,
  max_depth,
  no_cores = NULL,
  unique = TRUE,
  seed = NULL,
  save = TRUE,
  file_name = NULL
)
}
\arguments{
\item{n}{Number of functions to be sampled from grammar}

\item{grammar}{A Grammar Object, created by the create_grammar() function.}

\item{max_depth}{Maximum recursive depth used to sample grammar.}

\item{no_cores}{The number of cores for parallel computation. If NULL than all but 2 cores are used.}

\item{unique}{Should only the uniquely sampled functions be kept.}

\item{seed}{An integer to be supplied to set.seed, or NULL not to set reproducible seeds.}

\item{save}{if a .feather of the output should be saved in the current working directory.}
}
\value{
Returns a data frame with all sampled functions.
}
\description{
grammar_sampler
}
\examples{
simple_grammar <- create_grammar(a = "<b><op><c>, <a><op><b>, <a><op><c>, 1",
                                 b = "2, 4",
                                 c = "1, 3, 5",
                                 op = "+, -")
\dontrun{
grammar_functions <- grammar_sampler(n = 10,
                                     grammar = simple_grammar,
                                     max_depth = 5,
                                     no_cores = 1,
                                     save = FALSE)
}
}
