% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_functions.R,
%   R/simplify_functions_WITH_BATCHES.R
\name{simplify_functions}
\alias{simplify_functions}
\title{simplify_functions}
\usage{
simplify_functions(files, batches = 10, no_cores)

simplify_functions(files, batches = 10, no_cores)
}
\arguments{
\item{files}{File paths to fst files containing functions that should be simplified.}

\item{no_cores}{Number of cores for parallel computations.}
}
\value{



}
\description{
simplify_functions

simplify_functions
}
