% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_functions.R
\name{simplify_functions}
\alias{simplify_functions}
\title{simplify_functions}
\usage{
simplify_functions(function_strings, function_variables, no_cores = NULL)
}
\arguments{
\item{function_variables}{Variables that are present in the functions. Function generated without any of the given variables will be returned as NAs.}

\item{no_cores}{Number of cores to use. If not defined, detectCores()-2 is used.}

\item{funs}{Character vector of functions to simplify.}
}
\value{
A character vector with simplified functions
}
\description{
This function is a combination of the simplify function of the package deriv and a method to include undefined numerics in a function and still be simplify the function correctly. The undefined numerics should be given in the function as "numeric". They will not be handled as 1 variable, but as many differnt ones. Therefore they will only be simplifyed in the two cases.
In case of multiple additions/substracion of only a numeric, e.g. +numeric+numeric-numeric., the will be summarazied into one numeric: +numeric-. In case of  more than one multiplication/substraction they will also be summarized, e.g. numeric*numeric/numeric -> numeric
}
